Refactor UserGeometry into its own file
Move UserGeometry out of GeoEngine.
