+++ app/src/main/java/org/scottishtecharmy/soundscape/geoengine/GeoEngine.kt
@@ -74,114 +74,7 @@ class GeoEngine {
 
     private val streetPreview = StreetPreview()
 
-    /** UserGeometry contains all of the data relating to the location and motion of the user. It's
-     * aim is to reduces the number of arguments to many of the API calls and to concentrate some of
-     * the logic around heading choice.
-     *
-     * @param phoneHeading is the direction in which the phone is pointing
-     * @param travelHeading is the direction in which the phone is moving
-     * @param headHeading is the direction in which the head tracking is pointing
-     *
-     * On iOS there were two types of heading prioritization:
-     *  collection: course (travel?), user (phone?), device (head?)
-     *  presentation:user (phone?), course (travel?), device (head?)
-     *
-     * It's very hard to follow the Heading code, but because head tracking isn't always (usually)
-     * present and I think `user` must be the phone direction and `device` the head tracking
-     * direction.
-     *
-     *  presentationHeading is used for audio beacons. This makes sense - it's the direction of the
-     *  phone that is used, though I'm surprised it ever uses directly of travel.
-     *  collectionHeading is used for intersections - the direction of travel is most significant
-     *  here.
-     *
-     * However, if the user has thrown their phone into their bag, we need to detect this and ignore
-     * the phone direction.
-     */
-    enum class HeadingMode {
-        Auto,
-        Phone,
-        Travel
-    }
-
-    class UserGeometry(val location: LngLatAlt = LngLatAlt(),
-                       var phoneHeading: Double = 0.0,
-                       val fovDistance: Double = 50.0,
-                       val inVehicle: Boolean = false,
-                       val inMotion: Boolean = false,
-                       val speed: Double = 0.0,
-                       private val headingMode: HeadingMode = HeadingMode.Auto,
-                       private var travelHeading: Double = Double.NaN,
-                       private var headHeading: Double = Double.NaN,
-                       private val inStreetPreview: Boolean = false)
-    {
-        private val automotiveRangeMultiplier = 6.0
-        private val streetPreviewRangeIncrement = 10.0
-
-        private fun transform(distance: Double) : Double {
-            if(inVehicle) return distance * automotiveRangeMultiplier
-            if(inStreetPreview) return distance + streetPreviewRangeIncrement
-            return distance
-        }
-
-        fun heading() : Double {
-            when(headingMode) {
-                HeadingMode.Auto -> {
-                    if(speed > 0.2 && !travelHeading.isNaN())
-                        return travelHeading
-
-                    return phoneHeading
-                }
-                HeadingMode.Phone -> return phoneHeading
-                HeadingMode.Travel -> return travelHeading
-            }
-        }
-
-        /**
-         * getSearchDistance returns the distance to use when searching for POIs
-         */
-        fun getSearchDistance() : Double {
-            return transform(50.0)
-        }
-
-        /**
-         * getTriggerRange returns the distance to use when detecting POIs to call out
-         */
-        fun getTriggerRange(category: String) : Double {
-            return when(category) {
-                "object",
-                "safety" -> transform(10.0)
-
-                "place",
-                "information",
-                "mobility" -> transform(20.0)
-
-                "landmark" -> transform(50.0)
-
-                else -> transform(0.0)
-            }
-        }
-
-        /**
-         * getTriggerRange returns the distance if a POI is still in proximity after a callout
-         */
-        fun getProximityRange(category: String) : Double {
-            return when(category) {
-                "object",
-                "safety" -> transform(20.0)
-
-                "place",
-                "information",
-                "mobility" -> transform(30.0)
-
-                "landmark" -> transform(100.0)
-
-                else -> transform(0.0)
-            }
-        }
-    }
-
-    private fun getCurrentUserGeometry(headingMode: HeadingMode) : UserGeometry {
+    private fun getCurrentUserGeometry(headingMode: UserGeometry.HeadingMode) : UserGeometry {
         return UserGeometry(
             location = locationProvider.get(),
             phoneHeading = directionProvider.getCurrentDirection().toDouble(),
@@ -278,7 +171,7 @@ class GeoEngine {
                     // falling back to use the phone direction.
                     if(!soundscapeService.isAudioEngineBusy()) {
                         val callouts =
-                            autoCallout.updateLocation(getCurrentUserGeometry(HeadingMode.Auto), gridState)
+                            autoCallout.updateLocation(getCurrentUserGeometry(UserGeometry.HeadingMode.Auto), gridState)
                         if (callouts.isNotEmpty()) {
                             // Tell the service that we've got some callouts to tell the user about
                             soundscapeService.speakCallout(callouts)
@@ -387,7 +280,7 @@ class GeoEngine {
             results = runBlocking {
                 withContext(gridState.treeContext) {
 
-                    val userGeometry = getCurrentUserGeometry(HeadingMode.Phone)
+                    val userGeometry = getCurrentUserGeometry(UserGeometry.HeadingMode.Phone)
 
                     // Direction order is: behind(0) left(1) ahead(2) right(3)
                     val featuresByDirection: Array<Feature?> = arrayOfNulls(4)
@@ -485,7 +378,7 @@ class GeoEngine {
                     val list: MutableList<PositionedString> = mutableListOf()
 
                     // get device direction
-                    val userGeometry = getCurrentUserGeometry(HeadingMode.Phone)
+                    val userGeometry = getCurrentUserGeometry(UserGeometry.HeadingMode.Phone)
 
                     // Detect if there is a road or an intersection in the FOV
                     val roadsDescription = getRoadsDescriptionFromFov(
@@ -569,7 +462,7 @@ class GeoEngine {
         val engine = this
         CoroutineScope(Job()).launch(gridState.treeContext) {
             // Get our current location and figure out what GO means
-            val userGeometry = getCurrentUserGeometry(HeadingMode.Phone)
+            val userGeometry = getCurrentUserGeometry(UserGeometry.HeadingMode.Phone)
             val choices = streetPreview.getDirectionChoices(engine, userGeometry.location)
             var heading = 0.0
             if(choices.isNotEmpty()) {
@@ -612,7 +505,7 @@ class GeoEngine {
         val engine = this
         CoroutineScope(Job()).launch(gridState.treeContext) {
             // Get our current location and figure out what GO means
-            val userGeometry = getCurrentUserGeometry(HeadingMode.Phone)
+            val userGeometry = getCurrentUserGeometry(UserGeometry.HeadingMode.Phone)
             streetPreview.go(userGeometry, engine)
         }
     }
