+++ app/src/test/java/org/scottishtecharmy/soundscape/IntersectionsTest.kt
@@ -4,8 +4,8 @@ import org.scottishtecharmy.soundscape.geojsonparser.geojson.FeatureCollection
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
 import org.junit.Assert
 import org.junit.Test
-import org.scottishtecharmy.soundscape.geoengine.GeoEngine
 import org.scottishtecharmy.soundscape.geoengine.GridState
+import org.scottishtecharmy.soundscape.geoengine.UserGeometry
 import org.scottishtecharmy.soundscape.geoengine.callouts.ComplexIntersectionApproach
 import org.scottishtecharmy.soundscape.geoengine.callouts.getRoadsDescriptionFromFov
 
@@ -18,7 +18,7 @@ class IntersectionsTest {
                           fovDistance: Double) : FeatureCollection {
 
         val gridState = GridState.createFromGeoJson(geoJsonResource)
-        val userGeometry = GeoEngine.UserGeometry(currentLocation, deviceHeading, fovDistance)
+        val userGeometry = UserGeometry(currentLocation, deviceHeading, fovDistance)
         return  getRoadsDescriptionFromFov(
                     gridState,
                     userGeometry,
