+++ app/src/test/java/org/scottishtecharmy/soundscape/filters/CalloutHistoryTest.kt
@@ -3,9 +3,8 @@ package org.scottishtecharmy.soundscape.filters
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
-import org.junit.Before
 import org.junit.Test
-import org.scottishtecharmy.soundscape.geoengine.GeoEngine
+import org.scottishtecharmy.soundscape.geoengine.UserGeometry
 import org.scottishtecharmy.soundscape.geoengine.filters.CalloutHistory
 import org.scottishtecharmy.soundscape.geoengine.filters.TrackedCallout
 import org.scottishtecharmy.soundscape.geojsonparser.geojson.LngLatAlt
@@ -34,7 +33,7 @@ class CalloutHistoryTest {
         )
         history.add(callout)
         Thread.sleep(20)
-        history.trim(GeoEngine.UserGeometry(location))
+        history.trim(UserGeometry(location))
         assertFalse(history.find(callout))
         assertEquals(0, history.size())
     }
@@ -46,7 +45,7 @@ class CalloutHistoryTest {
         val distantLocation = LngLatAlt(10.0, 10.0, 10.0)
         val callout = TrackedCallout("Poi number three", distantLocation, true, false)
         history.add(callout)
-        history.trim(GeoEngine.UserGeometry(location))
+        history.trim(UserGeometry(location))
         assertFalse(history.find(callout))
         assertEquals(0, history.size())
     }
@@ -55,7 +54,7 @@ class CalloutHistoryTest {
     fun testTrimEmptyHistory() {
         val history = CalloutHistory(10)
         val location = LngLatAlt(0.0, 0.0, 0.0) // Example location
-        history.trim(GeoEngine.UserGeometry(location))
+        history.trim(UserGeometry(location))
         assertEquals(0, history.size())
     }
 }
